Skip test if a module is missing and use Test::More
Rather than generating a failure when an optional module is missing
we simply report the fact and skip the test. Compilation errors
are still a failure.

For some reason Test::Harness does not report skip messages. I
has assumed this was a problem with the manual message generation
but switching to Test::More did not help. Note that Test::More
required the tests to be reported by the parent process so now
the forks write the result to a temp file and pass it back
to the parent for reporting.

Additionally the compilation errors are reported using the
diag() function.