Specify catalog location in SkyCat GSC test
This should prevent the test 4_skycat_gsc.t from failing by
selecting a random location for the catalog.  The comments
in the test script indicate that ESO was the intended source.
The test script fails if CADC is chosen as the source because
the constructed URL retrieves the data in HTML format rather
than text.