RITMatch: Fix case with multiple filter choices
Astro::Fluxes stores the filters in a hash which means that the "first"
filter is probably going to be different for each star as they are written
out. This leads to some strange catalogues being written. Additionally
some filters have missing values and these were not being properly trapped.

Now ensure that the filter selected for the first star is used for
all subsequent entries. Also allow a particular filter to be specified
as an argument.

Finally, now that the catalogue no longer contains gaps due to missing
entries we can now also write the original star name in an extra column.