2_jcmt.t: Skip HOLO source distance check
The distance is calculated in RA / Dec coordinates, where the
HOLO source moves.  So if there is a delay between the conversion
of the two sets of coordinates to RA / Dec, there can be
a distance between them causing the test to fail.  Therefore
skip this test for the HOLO source.