bin: Add "catalog_convert" script
This script can be used to convert between catalog formats known
to this module.  It does so by reading a catalog (using one of the
Astro::Catalog::IO modules) and then re-writing it (using another of
the Astro::Catalog::IO modules).  If it can not guess the formats of
the catalogs from the file extensions then they must be specified via
command line options.