IO::JCMT: use per-source counter duplicate suffix
Using the unknown source counter results in quite large numbers
being generated with the current JCMT pointing catalog (if the
check for equal coordinates is disabled).  Removed the punctuation
from the end of the generated source name warnings so that this
is not mistaken for part of the new source name.