IO::JCMT: add inccomments read option
This option can be used to read comments from a catalog into the
"misc" hash of catalog items.  They are then included if the catalog
is written in JCMT format.  (In which case incheader should be disabled
to prevent addition of extra header comments.)

It would be nice to be able to include comments as individual entries
in the catalog, perhaps as a special class of Astro::Catalog::Item,
with everything updated to support that.  However for now they are
attached to the following item, or the preceding item in the case
of comments at the end of the file.