IO::JCMT: add proper motion and parallax columns
This adds new columns giving the (coordinate angle of) proper motion
in RA and proper motion in Dec, in milli-arcsec/yr, and the parallax
in milli-arcsec.  Coordinates are now obtained from the "ra2000" and
"dec2000" methods so that they correspond to epoch 2000.

The new columns have been placed at the end so that any un-updated
software ignores them as part of the comment.