IO::VOTable: attempt to make module more tolerant
If the definitions section is not found, assume J2000 for now?

Use more tolerant patterns to identify the "id", "ra" and "dec"
columns.  Also check if the units for "ra" and "dec" start "deg",
in which case specify the units for these values as "degrees"
rather than "sexagesimal".

It is hoped that this will allow the module to read catalogs
generated by the JCMT Query Tool, while still handling any
format which was previously supported.