Item: Process "Coords" first in configure method
Some of the IO classes rely on constructor arguments (such as "ID")
being passed on to the Astro::Coords object (given in the "Coords"
argument).  However since the arguments were processed from a hash
in an undefined order, the coords object may or may not have been
present to allow this to happen.

Therefore check for the "Coords" argument and process it before
passing the remaining arguments to the corresponding accessor
methods.