IO::JCMT: correct handling of redshifts
The JCMT pointing catalog description does say that the velocity
column is in km/s, but where the definition is redshift, an actual
redshift value would make more sense.  Therefore get the value
with "->redshift" instead of "->rv" in this case and format it
with more decimal places.  When reading a redshift, pass it to
Astro::Coords as such (rather than as a velocity in km/s).
Truncate the velocity definition to 5 characters to match the
current pointing catalog (where every entry currently says
"RADIO") in order to maintain alignment of the columns.
(Note that Astro::Coords only inspects the first 3 characters
anyway when creating the coordinates object.)