Query::SIMBAD: begin update to use script query
Initial update of the Astro::Catalog::Query::SIMBAD module to
use the SIMBAD script query interface.  For now only searching
by object (name) is implemented.

Proper motion and parallax are retrieved and included in the
constructed Astro::Coords objects (with proper motion in RA
converted to coordinate angle as opposed to the on-sky
angle "mu-ra*cos(dec)" used by SIMBAD).