Transport::REST: correct user agent string
Previously this used "Astro::Catalog::REST" as the product name
but the version taken from the (sub)class actually in use.
"Astro::Catalog::REST" doesn't exist, and while it could say
"Astro::Catalog::Transport::REST", that isn't the module for
which it obtains the version number.  Also the client host
and domain name do not seem to belong here.